% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div-anchors-package.R
\docType{package}
\name{div-anchors-package}
\alias{divanchors}
\alias{div-anchors-package}
\title{div-anchors: A Quarto Extension for Theorem Div Anchor Links}
\description{
A Quarto extension that adds visual URL anchor links to theorem and proof
divs for HTML output, similarly to the built-in anchor-sections option for
section headings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/div-anchors}
  \item Report bugs at \url{https://github.com/d-morrison/div-anchors/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu}

}
\keyword{internal}
